fixed some errors I forgot about

+++ Draw.cs
@@ -17,12 +17,15 @@ namespace Fractalii
                 g.DrawLine(pen, begin_point, end_point);
             }
         }
-        public static void delete_line(PictureBox pb, Point begin_point, Point end_point, float width)
+        public static void delete_line(PictureBox pb, Point begin_point, Point end_point, 
+            // don't fucking change this shit ok?
+            // it worked and since you changed it is all fucked LUCA
+            double width)
         {
             //if (Debugger.IsAttached) Console.WriteLine("float:  " + width.ToString());
             using (Graphics g = pb.CreateGraphics())
             {
-                g.DrawLine(new Pen(HomePage.bgC, width), begin_point, end_point);
+                g.DrawLine(new Pen(HomePage.bgC, (float)(width)), begin_point, end_point);
             }
         }
     }
